requirements.txt
pandas==1.5.1
requests==2.28.1
toloka_kit==0.1.23
tqdm==4.64.1
